feat(user): add service & controller

+++ src/main/java/com/example/demo/service/UserService.java
@@ -1,17 +1,19 @@
 package com.example.demo.service;
 
 import com.example.demo.entity.User;
+import com.example.demo.exception.AppException;
 import com.example.demo.repository.UserRepository;
 import com.example.demo.specification.GenericSpecification;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-@Service 
+@Service
 public class UserService {
+
   private final UserRepository userRepository;
 
   public UserService(UserRepository userRepository) {
@@ -27,22 +29,30 @@ public class UserService {
     return userRepository.findAll(GenericSpecification.searchByColumn(search, columnToSearch));
   }
 
-  public Optional<User> getUserById(String id) {
-    return  userRepository.findById(id);
+  public User getUserById(String id) {
+    return userRepository.findById(id)
+      .orElseThrow(() -> new AppException(HttpStatus.NOT_FOUND, "User with ID " + id + " was not found")
+    );
   }
 
   public User updateUser(String id, User userDetails) {
-    return userRepository.findById(id).map(user -> {
-      user.setRole(userDetails.getRole());
-      user.setStatus(userDetails.getStatus());
-      user.setPhoneNumber(userDetails.getPhoneNumber());
-      user.setProfileImageUrl(userDetails.getProfileImageUrl());
+    return userRepository.findById(id)
+      .map(user -> {
+        user.setRole(userDetails.getRole());
+        user.setStatus(userDetails.getStatus());
+        user.setPhoneNumber(userDetails.getPhoneNumber());
+        user.setProfileImageUrl(userDetails.getProfileImageUrl());
 
-      return userRepository.save(user);
-    }).orElseThrow(() -> new RuntimeException("User not found"));
+        return userRepository.save(user);
+      })
+      .orElseThrow(() -> new AppException(HttpStatus.NOT_FOUND, "User with ID " + id + " was not found")
+    );
   }
 
   public void deleteUser(String id) {
+    if (!userRepository.existsById(id)) {
+      throw new AppException(HttpStatus.NOT_FOUND, "User with ID " + id + " was not found");
+    }
     userRepository.deleteById(id);
   }
 }
